implemented pull request suggestions and consolidated annotation state and logic into new class
