+++ package.json
@@ -30,8 +30,8 @@
     "@types/dompurify": "^2.0.3",
     "@types/jest": "^25.1.4",
     "@types/loglevel": "^1.6.3",
-    "@types/marked": "^1.1.0",
     "@types/node": "^12.12.53",
+    "@types/marked": "^1.1.0",
     "@types/storejs": "^2.0.3",
     "@types/upng-js": "^2.1.1",
     "@types/uuid": "^8.3.0",
