+++ src/eterna/mode/PoseEdit/SubmitPoseDetails.ts
@@ -1,7 +1,7 @@
-import {AnnotationGraph} from 'eterna/ui/AnnotationItem';
+import {AnnotationDataBundle} from 'eterna/AnnotationManager';
 
 export default interface SubmitPoseDetails {
     title: string | undefined;
     comment: string | undefined;
-    annotations: AnnotationGraph;
+    annotations: AnnotationDataBundle;
 }
