+++ src/eterna/net/GameClient.ts
@@ -4,7 +4,7 @@ import {FoldData} from 'eterna/UndoBlock';
 import {VoteData} from 'eterna/mode/DesignBrowser/VoteProcessor';
 import {RankScrollData} from 'eterna/rank/RankScroll';
 import {PuzzleJSON} from 'eterna/puzzle/PuzzleManager';
-import {AnnotationGraph} from 'eterna/ui/AnnotationItem';
+import {AnnotationDataBundle} from 'eterna/AnnotationManager';
 
 // we MUST do this, because we use a dom method that gives us
 // any. long term AMW TODO build interfaces for each get/post
@@ -21,8 +21,8 @@ interface VoteDataWrapper {
 
 type Params =Record<
 string,
-string | number | boolean | PuzzleJSON | RankScrollData |
-AnnotationGraph | null | undefined>;
+string | number | boolean | PuzzleJSON | RankScrollData | AnnotationDataBundle | null | undefined
+>;
 
 export default class GameClient {
     public readonly baseURL: string;
