+++ src/eterna/puzzle/SolutionManager.ts
@@ -3,7 +3,7 @@ import EPars from 'eterna/EPars';
 import Eterna from 'eterna/Eterna';
 import Feedback, {BrentTheoData} from 'eterna/Feedback';
 import Sequence from 'eterna/rnatypes/Sequence';
-import {AnnotationGraph} from 'eterna/ui/AnnotationItem';
+import {AnnotationDataBundle} from 'eterna/AnnotationManager';
 import Solution from './Solution';
 
 interface SolutionSpec {
@@ -29,7 +29,7 @@ interface SolutionSpec {
     'SHAPE-min': string;
     'has-fold-data': number | null;
     'fold-data': string;
-    'annotations': AnnotationGraph;
+    'annotations': AnnotationDataBundle;
 }
 
 interface ShapeData {
@@ -254,7 +254,7 @@ export default class SolutionManager {
         }
 
         if (obj['annotations'] != null) {
-            newsol.annotations = obj['annotations'] as AnnotationGraph;
+            newsol.annotations = obj['annotations'] as AnnotationDataBundle;
         }
 
         return newsol;
