+++ src/eterna/ui/AnnotationPanelItem.ts
@@ -1,6 +1,9 @@
 import {UnitSignal, Value} from 'signals';
 import {
-    Container, Graphics, Sprite, Text, Point, Rectangle
+    Container,
+    Graphics,
+    Sprite,
+    Text
 } from 'pixi.js';
 import {
     ContainerObject,
@@ -13,130 +16,35 @@ import Bitmaps from 'eterna/resources/Bitmaps';
 import Fonts from 'eterna/util/Fonts';
 import TextBuilder, {FontWeight} from 'flashbang/util/TextBuilder';
 import Eterna from 'eterna/Eterna';
+import {
+    AnnotationData,
+    AnnotationHierarchyType,
+    AnnotationCategory,
+    AnnotationRange
+} from 'eterna/AnnotationManager';
 import GameButton from './GameButton';
 import DragDropper, {DragDropType, Item} from './DragDropper';
 import TextInputObject from './TextInputObject';
 import AnnotationDialog from './AnnotationDialog';
-import AnnotationCard from './AnnotationCard';
-
-export interface AnnotationData {
-    id: string | number;
-    type: AnnotationItemType;
-    category?: AnnotationCategory;
-    timestamp?: number;
-    playerID: number;
-    title: string;
-    ranges?: AnnotationRange[];
-    layer?: AnnotationData;
-    children?: AnnotationData[];
-    visible?: boolean;
-    selected?: boolean;
-}
-
-export interface AnnotationArguments {
-    ranges: AnnotationRange[];
-}
-
-export interface AnnotationDataCollection {
-    puzzle: AnnotationData[];
-    solution: AnnotationData[];
-}
-
-export interface AnnotationDisplayObject {
-    data: AnnotationData;
-    type: AnnotationItemType;
-    positions: AnnotationPosition[];
-    displays: AnnotationCard[];
-}
-
-export interface AnnotationPositionConflict {
-    bounds: Rectangle;
-    placement: AnnotationPlacement;
-}
-
-export interface AnnotationBaseConflict {
-    bounds: Rectangle;
-    resolvable: boolean;
-    correction?: Point;
-}
-
-export enum AnnotationPlacement {
-    TOP_LEFT = 'top-left',
-    TOP_CENTER = 'top-center',
-    TOP_RIGHT = 'top-right',
-    LEFT_CENTER = 'left-center',
-    RIGHT_CENTER = 'right-center',
-    BOTTOM_LEFT = 'bottom-left',
-    BOTTOM_CENTER = 'bottom-center',
-    BOTTOM_RIGHT = 'bottom-right',
-}
-
-export interface AnnotationBaseConflicts {
-    [AnnotationPlacement.TOP_LEFT]: AnnotationBaseConflict | null;
-    [AnnotationPlacement.TOP_CENTER]: AnnotationBaseConflict | null;
-    [AnnotationPlacement.TOP_RIGHT]: AnnotationBaseConflict | null;
-    [AnnotationPlacement.LEFT_CENTER]: AnnotationBaseConflict | null;
-    [AnnotationPlacement.RIGHT_CENTER]: AnnotationBaseConflict | null;
-    [AnnotationPlacement.BOTTOM_LEFT]: AnnotationBaseConflict | null;
-    [AnnotationPlacement.BOTTOM_CENTER]: AnnotationBaseConflict | null;
-    [AnnotationPlacement.BOTTOM_RIGHT]: AnnotationBaseConflict | null;
-}
-
-export interface AnnotationPosition {
-    anchorIndex: number;
-    zoomLevel: number;
-    relPosition: Point;
-    placement?: AnnotationPlacement;
-    custom: boolean;
-}
-
-export interface AnnotationGraphNode {
-    data: AnnotationData;
-    type: AnnotationItemType;
-    positions: AnnotationPosition[];
-    children: AnnotationGraphNode[];
-}
-
-export interface AnnotationGraph {
-    puzzle: AnnotationGraphNode[];
-    solution: AnnotationGraphNode[];
-}
-
-export interface AnnotationRange {
-    start: number;
-    end: number;
-}
-
-export enum AnnotationCategory {
-    STRUCTURE = 'Structure',
-    PUZZLE = 'Puzzle',
-    SOLUTION = 'Solution'
-}
-
-export enum AnnotationItemType {
-    CATEGORY = 'category',
-    LAYER = 'layer',
-    ANNOTATION = 'annotation'
-}
 
 interface TruncatedLabelText {
     object: Text;
     string: string;
 }
 
-interface AnnotationItemProps {
+interface AnnotationPanelItemProps {
     id: string | number;
     indexPath: number[];
     width: number;
     dividerThickness: number;
-    type: AnnotationItemType;
+    type: AnnotationHierarchyType;
     title: string;
     category: AnnotationCategory;
-    children?: AnnotationData[];
+    children: AnnotationData[];
     timestamp?: number;
     playerID: number;
     ranges?: AnnotationRange[];
-    layer?: AnnotationData;
+    layerId?: string | number;
     body?: string;
     titleEditable: boolean;
     updateTitle: (itemPath: number[], text: string) => void;
@@ -144,13 +52,13 @@ interface AnnotationItemProps {
     updateAnnotationPosition: (firstAnnotation: Item, secondAnnotationPath: number[]) => void;
 }
 
-export default class AnnotationItem extends ContainerObject {
+export default class AnnotationPanelItem extends ContainerObject {
     public readonly isVisible: Value<boolean> = new Value<boolean>(true);
     public readonly isSelected: Value<boolean> = new Value<boolean>(false);
     public readonly isExpanded: Value<boolean> = new Value<boolean>(true);
     public readonly isEditingTitle: Value<boolean> = new Value<boolean>(false);
     public readonly onUpdatePanel = new UnitSignal();
-    constructor(props: AnnotationItemProps) {
+    constructor(props: AnnotationPanelItemProps) {
         super();
         this._id = props.id;
         this._indexPath = props.indexPath;
@@ -169,8 +77,8 @@ export default class AnnotationItem extends ContainerObject {
         if (props.ranges) {
             this._ranges = props.ranges;
         }
-        if (props.layer) {
-            this._layer = props.layer;
+        if (props.layerId) {
+            this._layerId = props.layerId;
         }
         if (props.body) {
             this._body = props.body;
@@ -180,31 +88,19 @@ export default class AnnotationItem extends ContainerObject {
         this._updateAnnotationLayer = props.updateAnnotationLayer;
         this._updateAnnotationPosition = props.updateAnnotationPosition;
 
-        let layer;
-        if (this._type === AnnotationItemType.LAYER) {
-            layer = {
-                id: props.id,
-                type: props.type,
-                title: props.title,
-                children: props.children,
-                playerID: props.playerID,
-                timestamp: props.timestamp
-            };
-        }
-
-        if (props.children) {
+        if (props.children.length > 0) {
             // Lay out annotation item children
             let numAnnotationChildren = 0;
-            const itemChildren: AnnotationItem[] = [];
+            const modelChildren: AnnotationPanelItem[] = [];
             for (let i = 0; i < props.children.length; i++) {
                 const child = props.children[i];
-                if (child.type === AnnotationItemType.LAYER) {
+                if (child.type === AnnotationHierarchyType.LAYER) {
                     numAnnotationChildren += child.children?.length || 0;
-                } else if (child.type === AnnotationItemType.ANNOTATION) {
+                } else if (child.type === AnnotationHierarchyType.ANNOTATION) {
                     numAnnotationChildren += 1;
                 }
 
-                const item = new AnnotationItem({
+                const item = new AnnotationPanelItem({
                     id: child.id,
                     indexPath: [...this._indexPath, i],
                     width: this._width,
@@ -216,7 +112,7 @@ export default class AnnotationItem extends ContainerObject {
                     timestamp: child.timestamp,
                     playerID: child.playerID,
                     ranges: child.ranges,
-                    layer,
+                    layerId: child.layerId,
                     titleEditable: this._titleEditable && child.playerID === Eterna.playerID,
                     updateTitle: this._updateTitle,
                     updateAnnotationLayer: this._updateAnnotationLayer,
@@ -236,7 +132,7 @@ export default class AnnotationItem extends ContainerObject {
                     // Adds back children:
                     // 1) child layers (if item is a category header)
                     // 2) child annotations (if item is a layer)
-                    for (const itemChild of this._itemChildren) {
+                    for (const itemChild of this._modelChildren) {
                         this._itemStack.addChild(itemChild.container);
                     }
 
@@ -251,10 +147,10 @@ export default class AnnotationItem extends ContainerObject {
                     item.isVisible.value = visible;
                 });
 
-                itemChildren.push(item);
+                modelChildren.push(item);
             }
             this._numAnnotationChildren = numAnnotationChildren;
-            this._itemChildren = itemChildren;
+            this._modelChildren = modelChildren;
         }
 
         this._itemStack = new VLayoutContainer(0, HAlign.CENTER);
@@ -267,47 +163,47 @@ export default class AnnotationItem extends ContainerObject {
         super.added();
         let length: number;
         switch (this._type) {
-            case AnnotationItemType.CATEGORY:
-                length = AnnotationItem.CATEGORY_HEIGHT;
+            case AnnotationHierarchyType.CATEGORY:
+                length = AnnotationPanelItem.CATEGORY_HEIGHT;
                 break;
-            case AnnotationItemType.LAYER:
-                length = AnnotationItem.LAYER_HEIGHT;
+            case AnnotationHierarchyType.LAYER:
+                length = AnnotationPanelItem.LAYER_HEIGHT;
                 break;
             default:
-                length = AnnotationItem.ANNOTATION_HEIGHT;
+                length = AnnotationPanelItem.ANNOTATION_HEIGHT;
                 break;
         }
 
         let usedWidth = 0; // tracks how much horizontal space is being filled as we build out item container
         if (
-            this._type === AnnotationItemType.LAYER
-            || (this._type === AnnotationItemType.ANNOTATION && this._layer === undefined)
+            this._type === AnnotationHierarchyType.LAYER
+            || (this._type === AnnotationHierarchyType.ANNOTATION && this._layerId === undefined)
         ) {
             // Layers or layerless annotations have a single indent
-            const width = AnnotationItem.LAYER_HEIGHT;
+            const width = AnnotationPanelItem.LAYER_HEIGHT;
             this._itemIndent = new Graphics()
-                .beginFill(AnnotationItem.ITEM_BACKGROUND_RESTING)
+                .beginFill(AnnotationPanelItem.ITEM_BACKGROUND_RESTING)
                 .drawRect(
                     0,
                     0,
                     width,
-                    this._type === AnnotationItemType.LAYER
-                        ? AnnotationItem.LAYER_HEIGHT
-                        : AnnotationItem.ANNOTATION_HEIGHT
+                    this._type === AnnotationHierarchyType.LAYER
+                        ? AnnotationPanelItem.LAYER_HEIGHT
+                        : AnnotationPanelItem.ANNOTATION_HEIGHT
                 )
                 .endFill();
             usedWidth += width;
             this._itemContainer.addChild(this._itemIndent);
-        } else if (this._type === AnnotationItemType.ANNOTATION && this._layer) {
+        } else if (this._type === AnnotationHierarchyType.ANNOTATION && this._layerId) {
             // Annotations in a layer have a double indent
-            const width = 2 * AnnotationItem.LAYER_HEIGHT;
+            const width = 2 * AnnotationPanelItem.LAYER_HEIGHT;
             this._itemIndent = new Graphics()
-                .beginFill(AnnotationItem.ITEM_BACKGROUND_RESTING)
+                .beginFill(AnnotationPanelItem.ITEM_BACKGROUND_RESTING)
                 .drawRect(
                     0,
                     0,
                     width,
-                    AnnotationItem.ANNOTATION_HEIGHT
+                    AnnotationPanelItem.ANNOTATION_HEIGHT
                 )
                 .endFill();
             usedWidth += width;
@@ -316,7 +212,7 @@ export default class AnnotationItem extends ContainerObject {
 
         // Set up visibility button
         this._visibilityButtonBackground = new Graphics()
-            .beginFill(AnnotationItem.VISIBILITY_BUTTON_BACKGROUND_COLOR)
+            .beginFill(AnnotationPanelItem.VISIBILITY_BUTTON_BACKGROUND_COLOR)
             .drawRect(
                 0,
                 0,
@@ -354,20 +250,20 @@ export default class AnnotationItem extends ContainerObject {
         let ribbonColor: number;
         if (this._category === AnnotationCategory.STRUCTURE) {
             // RNA Annotation or Layer
-            ribbonColor = AnnotationItem.STRUCTURE_RIBBON_COLOR;
+            ribbonColor = AnnotationPanelItem.STRUCTURE_RIBBON_COLOR;
         } else if (this._category === AnnotationCategory.PUZZLE) {
             // PLAYER Annotation or Layer
-            ribbonColor = AnnotationItem.PUZZLE_RIBBON_COLOR;
+            ribbonColor = AnnotationPanelItem.PUZZLE_RIBBON_COLOR;
         } else {
             // SOLUTION Annotation or Layer
-            ribbonColor = AnnotationItem.SOLUTION_RIBBON_COLOR;
+            ribbonColor = AnnotationPanelItem.SOLUTION_RIBBON_COLOR;
         }
         this._itemRibbon = new Graphics()
             .beginFill(ribbonColor)
             .drawRect(
                 0,
                 0,
-                AnnotationItem.RIBBON_WIDTH,
+                AnnotationPanelItem.RIBBON_WIDTH,
                 length
             )
             .endFill();
@@ -375,8 +271,8 @@ export default class AnnotationItem extends ContainerObject {
 
         // Set up accordion chevron if necessary
         if (
-            (this._type === AnnotationItemType.CATEGORY && this._itemChildren && this._itemChildren.length > 0)
-            || (this._type === AnnotationItemType.LAYER && this._itemChildren && this._itemChildren.length > 0)
+            (this._type === AnnotationHierarchyType.CATEGORY && this._modelChildren && this._modelChildren.length > 0)
+            || (this._type === AnnotationHierarchyType.LAYER && this._modelChildren && this._modelChildren.length > 0)
         ) {
             const accordionChevronContainer = new Container();
             this._chevronRight = Sprite.from(Bitmaps.ImgChevronRight);
@@ -384,7 +280,7 @@ export default class AnnotationItem extends ContainerObject {
             this._chevronDown.alpha = 0;
             const aspectRatio = this._chevronRight.width / this._chevronRight.height;
             this._accordionChevron = new Graphics()
-                .beginFill(AnnotationItem.ITEM_BACKGROUND_RESTING)
+                .beginFill(AnnotationPanelItem.ITEM_BACKGROUND_RESTING)
                 .drawRect(
                     0,
                     0,
@@ -402,7 +298,7 @@ export default class AnnotationItem extends ContainerObject {
                     this._chevronDown.alpha = 1;
 
                     // add item children
-                    for (const child of this._itemChildren) {
+                    for (const child of this._modelChildren) {
                         this._itemStack.addChild(child.container);
                     }
 
@@ -412,9 +308,9 @@ export default class AnnotationItem extends ContainerObject {
                     this._chevronRight.alpha = 1;
                     this._chevronDown.alpha = 0;
 
-                    // remove item children
-                    this._itemStack.removeChildren();
-                    this._itemStack.addChild(this._itemContainer);
+                    // We want to keep the header but remove the children, hence starting
+                    // at index 1
+                    this._itemStack.removeChildren(1, this._itemStack.children.length);
 
                     // Updates vertical layout of annotation item + children
                     this._itemStack.layout();
@@ -422,7 +318,7 @@ export default class AnnotationItem extends ContainerObject {
 
                 this.isExpanded.value = isExpanded;
 
-                if (this._type === AnnotationItemType.CATEGORY) {
+                if (this._type === AnnotationHierarchyType.CATEGORY) {
                     // Notify that annotation layer panel that needs to be updated
                     // if we expand or close category accordion
                     this.onUpdatePanel.emit();
@@ -441,9 +337,9 @@ export default class AnnotationItem extends ContainerObject {
 
         // Set up layer name
         const textContainer = new Container();
-        this._itemWidth = this._width - length - AnnotationItem.RIBBON_WIDTH - usedWidth;
+        this._itemWidth = this._width - length - AnnotationPanelItem.RIBBON_WIDTH - usedWidth;
         this._itemButtonBackground = new Graphics()
-            .beginFill(AnnotationItem.ITEM_BACKGROUND_RESTING)
+            .beginFill(AnnotationPanelItem.ITEM_BACKGROUND_RESTING)
             .drawRect(
                 0,
                 0,
@@ -468,11 +364,11 @@ export default class AnnotationItem extends ContainerObject {
         const labelText = preparedLabelText.object;
         const truncatedText = preparedLabelText.string;
 
-        if (this._type === AnnotationItemType.CATEGORY) {
-            // Create immutable label
+        if (this._type === AnnotationHierarchyType.CATEGORY) {
+            // Create uneditable label
             labelText.y = (length - labelText.height) / 2;
-            if (!this._itemChildren || this._itemChildren.length === 0) {
-                labelText.x = AnnotationItem.ITEM_BUTTON_MARGIN_LEFT;
+            if (!this._modelChildren || this._modelChildren.length === 0) {
+                labelText.x = AnnotationPanelItem.ITEM_BUTTON_MARGIN_LEFT;
             }
             textContainer.addChild(labelText);
             this._itemButton.tooltip(this._title);
@@ -480,7 +376,7 @@ export default class AnnotationItem extends ContainerObject {
             // Create editabled label (click to reveal input)
             const itemTextButtonTextBuilder = new TextBuilder(truncatedText)
                 .font(Fonts.STDFONT)
-                .fontSize(AnnotationItem.FONT_SIZE)
+                .fontSize(AnnotationPanelItem.FONT_SIZE)
                 .fontWeight(FontWeight.REGULAR)
                 .color(0xFFFFFF)
                 .hAlignLeft();
@@ -525,19 +421,22 @@ export default class AnnotationItem extends ContainerObject {
 
             this._itemTextButton.display.y = (length - labelText.height) / 2;
             if (
-                this._type === AnnotationItemType.ANNOTATION
-                || (this._type === AnnotationItemType.LAYER && (!this._itemChildren || this._itemChildren.length === 0))
+                this._type === AnnotationHierarchyType.ANNOTATION
+                || (
+                    this._type === AnnotationHierarchyType.LAYER
+                    && (!this._modelChildren || this._modelChildren.length === 0)
+                )
             ) {
-                this._itemTextButton.display.x = AnnotationItem.ITEM_BUTTON_MARGIN_LEFT;
+                this._itemTextButton.display.x = AnnotationPanelItem.ITEM_BUTTON_MARGIN_LEFT;
             }
 
             this.addObject(this._itemTextButton, textContainer);
 
             // Create input
             this._itemNameInput = new TextInputObject({
-                fontSize: AnnotationItem.INPUT_FONT_SIZE,
-                width: this._itemWidth - 2 * AnnotationItem.INPUT_MARGIN,
-                height: length - 2 * AnnotationItem.INPUT_MARGIN,
+                fontSize: AnnotationPanelItem.INPUT_FONT_SIZE,
+                width: this._itemWidth - 2 * AnnotationPanelItem.INPUT_MARGIN,
+                height: length - 2 * AnnotationPanelItem.INPUT_MARGIN,
                 rows: 1,
                 placeholder: this._title,
                 characterLimit: AnnotationDialog.ANNOTATION_TEXT_CHARACTER_LIMIT
@@ -559,7 +458,7 @@ export default class AnnotationItem extends ContainerObject {
                         const truncText = editedPreparedLabelText.string;
                         const textBuilder = new TextBuilder(truncText)
                             .font(Fonts.STDFONT)
-                            .fontSize(AnnotationItem.FONT_SIZE)
+                            .fontSize(AnnotationPanelItem.FONT_SIZE)
                             .fontWeight(FontWeight.REGULAR)
                             .color(0xFFFFFF)
                             .hAlignLeft();
@@ -585,8 +484,8 @@ export default class AnnotationItem extends ContainerObject {
                 }
             });
 
-            this._itemNameInput.display.x = AnnotationItem.INPUT_MARGIN;
-            this._itemNameInput.display.y = AnnotationItem.INPUT_MARGIN;
+            this._itemNameInput.display.x = AnnotationPanelItem.INPUT_MARGIN;
+            this._itemNameInput.display.y = AnnotationPanelItem.INPUT_MARGIN;
             this._itemNameInput.display.visible = false;
 
             this.addObject(this._itemNameInput, textContainer);
@@ -613,7 +512,7 @@ export default class AnnotationItem extends ContainerObject {
 
         // Attach drag sources and drop targets
         switch (this._type) {
-            case AnnotationItemType.CATEGORY:
+            case AnnotationHierarchyType.CATEGORY:
                 this._dropTarget = new DragDropper({
                     dropType: DragDropType.TARGET,
                     itemDisplayObject: this.display,
@@ -623,13 +522,13 @@ export default class AnnotationItem extends ContainerObject {
                         index: this._indexPath,
                         type: this._type.toString()
                     },
-                    acceptItemType: AnnotationItemType.ANNOTATION.toString(),
+                    acceptItemType: AnnotationHierarchyType.ANNOTATION.toString(),
                     onDrop: (item: Item): void => this.handleDrop(item),
                     onHoverStart: (): void => this.activateHoverHighlight(),
                     onHoverEnd: (): void => this.deactivateHoverHighlight()
                 });
                 break;
-            case AnnotationItemType.LAYER:
+            case AnnotationHierarchyType.LAYER:
                 this._dragSource = new DragDropper({
                     dropType: DragDropType.SOURCE,
                     itemDisplayObject: this.display,
@@ -639,7 +538,7 @@ export default class AnnotationItem extends ContainerObject {
                         index: this._indexPath,
                         type: this._type.toString()
                     },
-                    draggingBorderColor: AnnotationItem.ITEM_HOVER_COLOR,
+                    draggingBorderColor: AnnotationPanelItem.ITEM_HOVER_COLOR,
                     canDrop: (): boolean => true
                 });
                 this._dropTarget = new DragDropper({
@@ -651,7 +550,7 @@ export default class AnnotationItem extends ContainerObject {
                         index: this._indexPath,
                         type: this._type.toString()
                     },
-                    acceptItemType: AnnotationItemType.ANNOTATION.toString(),
+                    acceptItemType: AnnotationHierarchyType.ANNOTATION.toString(),
                     onDrop: (item: Item): void => this.handleDrop(item),
                     onHoverStart: (): void => this.activateHoverHighlight(),
                     onHoverEnd: (): void => this.deactivateHoverHighlight()
@@ -668,7 +567,7 @@ export default class AnnotationItem extends ContainerObject {
                         index: this._indexPath,
                         type: this._type.toString()
                     },
-                    draggingBorderColor: AnnotationItem.ITEM_HOVER_COLOR,
+                    draggingBorderColor: AnnotationPanelItem.ITEM_HOVER_COLOR,
                     canDrop: (): boolean => true
                 });
                 this._dropTarget = new DragDropper({
@@ -680,7 +579,7 @@ export default class AnnotationItem extends ContainerObject {
                         index: this._indexPath,
                         type: this._type.toString()
                     },
-                    acceptItemType: AnnotationItemType.ANNOTATION.toString(),
+                    acceptItemType: AnnotationHierarchyType.ANNOTATION.toString(),
                     onDrop: (item: Item): void => this.handleDrop(item),
                     onHoverStart: (): void => this.activateHoverHighlight(),
                     onHoverEnd: (): void => this.deactivateHoverHighlight()
@@ -696,12 +595,12 @@ export default class AnnotationItem extends ContainerObject {
         }
 
         // Add any children to item stack
-        if (this._itemChildren && this._itemChildren.length > 0) {
+        if (this._modelChildren && this._modelChildren.length > 0) {
             this.isExpanded.value = true;
             this._chevronRight.alpha = 0;
             this._chevronDown.alpha = 1;
 
-            for (const child of this._itemChildren) {
+            for (const child of this._modelChildren) {
                 this.addObject(child, this._itemStack);
             }
 
@@ -714,7 +613,7 @@ export default class AnnotationItem extends ContainerObject {
             ? `${text} (${this._numAnnotationChildren})` : text;
         let labelText = new TextBuilder(fullText)
             .font(Fonts.STDFONT)
-            .fontSize(AnnotationItem.FONT_SIZE)
+            .fontSize(AnnotationPanelItem.FONT_SIZE)
             .fontWeight(FontWeight.REGULAR)
             .color(0xFFFFFF)
             .hAlignLeft()
@@ -728,7 +627,7 @@ export default class AnnotationItem extends ContainerObject {
                 ? `${truncatedText} (${this._numAnnotationChildren})` : truncatedText;
             labelText = new TextBuilder(fullText)
                 .font(Fonts.STDFONT)
-                .fontSize(AnnotationItem.FONT_SIZE)
+                .fontSize(AnnotationPanelItem.FONT_SIZE)
                 .fontWeight(FontWeight.REGULAR)
                 .color(0xFFFFFF)
                 .hAlignLeft()
@@ -741,7 +640,7 @@ export default class AnnotationItem extends ContainerObject {
                 : `${truncatedText.substring(0, truncatedText.length - 3)}...`;
             labelText = new TextBuilder(truncatedText)
                 .font(Fonts.STDFONT)
-                .fontSize(AnnotationItem.FONT_SIZE)
+                .fontSize(AnnotationPanelItem.FONT_SIZE)
                 .fontWeight(FontWeight.REGULAR)
                 .color(0xFFFFFF)
                 .hAlignLeft()
@@ -763,7 +662,7 @@ export default class AnnotationItem extends ContainerObject {
         if (this._itemIndent) {
             this.redraw(
                 this._itemIndent,
-                AnnotationItem.ITEM_BACKGROUND_SELECTED,
+                AnnotationPanelItem.ITEM_BACKGROUND_SELECTED,
                 this._itemIndent.width,
                 this._itemIndent.height
             );
@@ -771,14 +670,14 @@ export default class AnnotationItem extends ContainerObject {
         if (this._accordionChevron) {
             this.redraw(
                 this._accordionChevron,
-                AnnotationItem.ITEM_BACKGROUND_SELECTED,
+                AnnotationPanelItem.ITEM_BACKGROUND_SELECTED,
                 this._accordionChevron.width,
                 this._accordionChevron.height
             );
         }
         this.redraw(
             this._itemButtonBackground,
-            AnnotationItem.ITEM_BACKGROUND_SELECTED,
+            AnnotationPanelItem.ITEM_BACKGROUND_SELECTED,
             this._itemButtonBackground.width,
             this._itemButtonBackground.height
         );
@@ -792,7 +691,7 @@ export default class AnnotationItem extends ContainerObject {
         if (this._itemIndent) {
             this.redraw(
                 this._itemIndent,
-                AnnotationItem.ITEM_BACKGROUND_RESTING,
+                AnnotationPanelItem.ITEM_BACKGROUND_RESTING,
                 this._itemIndent.width,
                 this._itemIndent.height
             );
@@ -800,14 +699,14 @@ export default class AnnotationItem extends ContainerObject {
         if (this._accordionChevron) {
             this.redraw(
                 this._accordionChevron,
-                AnnotationItem.ITEM_BACKGROUND_RESTING,
+                AnnotationPanelItem.ITEM_BACKGROUND_RESTING,
                 this._accordionChevron.width,
                 this._accordionChevron.height
             );
         }
         this.redraw(
             this._itemButtonBackground,
-            AnnotationItem.ITEM_BACKGROUND_RESTING,
+            AnnotationPanelItem.ITEM_BACKGROUND_RESTING,
             this._itemButtonBackground.width,
             this._itemButtonBackground.height
         );
@@ -820,7 +719,7 @@ export default class AnnotationItem extends ContainerObject {
         const truncText = preparedLabelText.string;
         const textBuilder = new TextBuilder(truncText)
             .font(Fonts.STDFONT)
-            .fontSize(AnnotationItem.FONT_SIZE)
+            .fontSize(AnnotationPanelItem.FONT_SIZE)
             .fontWeight(FontWeight.REGULAR)
             .color(0xFFFFFF)
             .hAlignLeft();
@@ -872,7 +771,7 @@ export default class AnnotationItem extends ContainerObject {
         if (this._itemIndent) {
             this.redraw(
                 this._itemIndent,
-                AnnotationItem.ITEM_HOVER_COLOR,
+                AnnotationPanelItem.ITEM_HOVER_COLOR,
                 this._itemIndent.width,
                 this._itemIndent.height
             );
@@ -880,14 +779,14 @@ export default class AnnotationItem extends ContainerObject {
         if (this._accordionChevron) {
             this.redraw(
                 this._accordionChevron,
-                AnnotationItem.ITEM_HOVER_COLOR,
+                AnnotationPanelItem.ITEM_HOVER_COLOR,
                 this._accordionChevron.width,
                 this._accordionChevron.height
             );
         }
         this.redraw(
             this._itemButtonBackground,
-            AnnotationItem.ITEM_HOVER_COLOR,
+            AnnotationPanelItem.ITEM_HOVER_COLOR,
             this._itemButtonBackground.width,
             this._itemButtonBackground.height
         );
@@ -897,7 +796,7 @@ export default class AnnotationItem extends ContainerObject {
         if (this._itemIndent) {
             this.redraw(
                 this._itemIndent,
-                AnnotationItem.ITEM_BACKGROUND_RESTING,
+                AnnotationPanelItem.ITEM_BACKGROUND_RESTING,
                 this._itemIndent.width,
                 this._itemIndent.height
             );
@@ -905,25 +804,25 @@ export default class AnnotationItem extends ContainerObject {
         if (this._accordionChevron) {
             this.redraw(
                 this._accordionChevron,
-                AnnotationItem.ITEM_BACKGROUND_RESTING,
+                AnnotationPanelItem.ITEM_BACKGROUND_RESTING,
                 this._accordionChevron.width,
                 this._accordionChevron.height
             );
         }
         this.redraw(
             this._itemButtonBackground,
-            AnnotationItem.ITEM_BACKGROUND_RESTING,
+            AnnotationPanelItem.ITEM_BACKGROUND_RESTING,
             this._itemButtonBackground.width,
             this._itemButtonBackground.height
         );
     }
 
     private handleDrop(item: Item) {
-        if (this._type === AnnotationItemType.CATEGORY) {
+        if (this._type === AnnotationHierarchyType.CATEGORY) {
             this._updateAnnotationLayer(item, this._indexPath);
-        } else if (this._type === AnnotationItemType.LAYER) {
+        } else if (this._type === AnnotationHierarchyType.LAYER) {
             this._updateAnnotationLayer(item, this._indexPath);
-        } else if (this._type === AnnotationItemType.ANNOTATION) {
+        } else if (this._type === AnnotationHierarchyType.ANNOTATION) {
             this._updateAnnotationPosition(item, this._indexPath);
         }
     }
@@ -936,8 +835,8 @@ export default class AnnotationItem extends ContainerObject {
         return this._category;
     }
 
-    public get itemChildren() {
-        return this._itemChildren;
+    public get modelChildren() {
+        return this._modelChildren;
     }
 
     public get title() {
@@ -964,8 +863,8 @@ export default class AnnotationItem extends ContainerObject {
         return this._body;
     }
 
-    public get layer() {
-        return this._layer;
+    public get layerId() {
+        return this._layerId;
     }
 
     public get dragSource() {
@@ -994,7 +893,7 @@ export default class AnnotationItem extends ContainerObject {
     private _title: string;
     private _width: number;
     private _dividerThickness: number;
-    private _type: AnnotationItemType;
+    private _type: AnnotationHierarchyType;
     private _category: AnnotationCategory;
     private _titleEditable: boolean;
     private _itemStack: VLayoutContainer;
@@ -1012,7 +911,7 @@ export default class AnnotationItem extends ContainerObject {
     private _itemButtonBackground: Graphics;
     private _itemTextButton: GameButton;
     private _itemNameInput: TextInputObject;
-    private _itemChildren: AnnotationItem[];
+    private _modelChildren: AnnotationPanelItem[];
     private _numAnnotationChildren: number | null = null;
     private _itemWidth: number;
     private _dropTarget: DragDropper;
@@ -1026,7 +925,7 @@ export default class AnnotationItem extends ContainerObject {
     private _timestamp: number;
     private _playerID: number;
     private _ranges: AnnotationRange[];
-    private _layer: AnnotationData;
+    private _layerId: string | number;
     private _body: string;
 
     private static readonly CATEGORY_HEIGHT = 40;
